refactor docstring helpers
